Adjusting ex02 based on pull request
Spitting up each class into its own file
