+++ exercises/Stephen_Smith/ex02_oo_basics/Debug/ex02_Ste.58CBB3CC.tlog/ex02_Stephen_Smith.lastbuildstate
@@ -1,0 +1,2 @@
+#TargetFrameworkVersion=v4.0:PlatformToolSet=v142:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit:WindowsTargetPlatformVersion=10.0
+Debug|Win32|C:\Users\lot53126\Documents\NewStarter\newstarter\exercises\Stephen_Smith\ex02_oo_basics\|
