+++ exercises/Stephen_Smith/ex02_oo_basics/ex02_Stephen_Smith.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29411.108
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ex02_Stephen_Smith", "ex02_Stephen_Smith.vcxproj", "{58CBB3CC-98FA-4E0B-A4EA-6139874E3B5B}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{58CBB3CC-98FA-4E0B-A4EA-6139874E3B5B}.Debug|x64.ActiveCfg = Debug|x64
+		{58CBB3CC-98FA-4E0B-A4EA-6139874E3B5B}.Debug|x64.Build.0 = Debug|x64
+		{58CBB3CC-98FA-4E0B-A4EA-6139874E3B5B}.Debug|x86.ActiveCfg = Debug|Win32
+		{58CBB3CC-98FA-4E0B-A4EA-6139874E3B5B}.Debug|x86.Build.0 = Debug|Win32
+		{58CBB3CC-98FA-4E0B-A4EA-6139874E3B5B}.Release|x64.ActiveCfg = Release|x64
+		{58CBB3CC-98FA-4E0B-A4EA-6139874E3B5B}.Release|x64.Build.0 = Release|x64
+		{58CBB3CC-98FA-4E0B-A4EA-6139874E3B5B}.Release|x86.ActiveCfg = Release|Win32
+		{58CBB3CC-98FA-4E0B-A4EA-6139874E3B5B}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {ABF3FF63-71E3-41E0-AA71-1920E548C304}
+	EndGlobalSection
+EndGlobal
