+++ exercises/Stephen_Smith/ex02_oo_basics/include/ShapeSorter.h
@@ -2,40 +2,33 @@
 #define SHAPESORTER_H_
 
 #include "Shapes.h"
+#include "Square.h"
+#include "Rectangle.h"
+#include "Triangle.h"
+#include "Circle.h"
 #include <algorithm>
-
-
+/**
+ * This class stores a number of shapes, and provides
+ * operations based on the number of sides, area, perimeter,
+ * and type of the stored shapes.
+ */
 class ShapeSorter
 {
-    public:
-
+public:
     // Constructors
     ShapeSorter();
-    ShapeSorter(const std::vector<std::shared_ptr<BaseShape>>& shapeList);
-
+    ShapeSorter(std::vector<BaseShape*> shapeList);
     // Add shapes to object
     void addShape();
-
     // Print functions
     void PrintShapeNames() const;
     void printShapeByType(const std::string& shapeType) const;
     void printShapeBySides(const int nSides) const ;
     void printShapesByArea() ;
     void printShapesByPerimeter();
-
- 
-    private:
-        // Vector of shapes sorted within the object
-        std::vector<std::shared_ptr<BaseShape>> m_shapeList;
+private:
+     // Vector of shapes sorted within the object
+     std::vector<BaseShape*> m_shapeList;
 };
-
-
-
-
-
-
-
-
-
 
 #endif
